+++ services/AnnotationsService.js
@@ -58,7 +58,25 @@ const readAnnotations = ({ assembly, chrom, start, end, family, nrph }) => new P
 
       // Obtain simple/tandem annotations from the TE_Idx
       let teidx_args = ["--assembly", full_assembly, "idx-query", "--data-type", "masks", "--chrom", chrom, "--start", start, "--end", end];
-      const tandem_annots = await te_idx.query(teidx_args);
+      let tandem_annots = await te_idx.query(teidx_args);
+      // RMH: 20251121: Added normalization step to convert strings to integers and remove whitespace.
+      //      The API requires that these fields be integers and this wasn't caught by our unit tests.
+      //      In addition there is a bug in a version of the te_idx tool (v0.1.1) which reports a 
+      //      line feed in the repeat_length field:
+      //           {
+      //             "repeat_length": "5\n",
+      //             "repeat_str": "GAGGG",
+      //             "seq_acc": "chr1",
+      //             "seq_end": "168146209",
+      //             "seq_start": "168146103"
+      //           },
+      // Normalize the results
+      tandem_annots = tandem_annots.map(entry => ({
+        ...entry,
+        repeat_length: parseInt(entry.repeat_length.replace(/\s+/g, ''), 10),
+        seq_start: parseInt(entry.seq_start, 10),
+        seq_end: parseInt(entry.seq_end, 10),
+      }));
 
       // Obtain the TE annotations from the TE_Idx
       teidx_args = ["--assembly", full_assembly, "idx-query", "--data-type", "assembly_alignments",  "--chrom", chrom, "--start", start, "--end", end];
@@ -72,7 +90,22 @@ const readAnnotations = ({ assembly, chrom, start, end, family, nrph }) => new P
         teidx_args.push("--nrph");
       }
 
-      const teResults = await te_idx.query(teidx_args);
+      let teResults = await te_idx.query(teidx_args);
+
+      // RMH: 20251121: Added normalization step to convert strings to integers and remove whitespace.
+      //      The API requires that these fields be integers and this wasn't caught by our unit tests.
+      //      Also the bit_score field may be a float.
+      // Normalize the results
+      teResults = teResults.map(entry => ({
+        ...entry,
+        bit_score: parseFloat(entry.bit_score),
+        ali_start: parseInt(entry.ali_start, 10),
+        ali_end: parseInt(entry.ali_end, 10),
+        model_start: parseInt(entry.model_start, 10),
+        model_end: parseInt(entry.model_end, 10),
+        seq_start: parseInt(entry.seq_start, 10),
+        seq_end: parseInt(entry.seq_end, 10),
+      }));
 
       // Collect all accessions found, as well as thier positions in the list of results
       // Fixed: This was defined as implicitly global (e.g. not 'let', 'const' or 'var').